feat(api): generalize crawler file uploads out of CSV-specific core code (#356)
routes/csvUploads.js was a core file, mounted unconditionally, that was
entirely CSV-specific — violating the "no crawler-specific code in core"
rule enforced everywhere else in the crawler plugin system (CrawlersPage.jsx
has zero hardcoded crawler-type literals; discover.js is loaded dynamically
per-type with no core file listing which crawlers have one). Traced this to
three places with hardcoded 'csv' logic, not just the one file:

- routes/csvUploads.js: folder-naming convention, config-type assertion,
  upload extension allowlist, route paths — all CSV-only.
- routes/jobs.js: a hardcoded `if (jobType === 'csv')` block injecting the
  upload folder into a job before dispatch.
- The wizard calling CSV-named URLs (fine — it's allowed to know it's CSV,
  just needed to follow the new generic route shape).

Renamed csvUploads.js -> routes/crawlerFiles.js and generalized it:
- Folder convention is now `{crawlerType}-{configId}`, gated by a new
  `supportsFileUploads` flag in crawler.json (kept this guard rather than
  dropping it — it's a real authorization-adjacent check, not just defensive
  padding: without it, any valid config id of any type could be used to
  create/list/delete files via this generic route).
- multer's extension allowlist now reads `uploadFileExtensions` from the
  manifest (falls back to ['.csv'] if a type opts into uploads without
  declaring extensions).
- Routes generalized: /crawler-configs/:id/csv-files -> .../files, and the
  schema-download endpoints moved to /admin/crawlers/:type/upload-schema
  (mirrors the discover.js per-type dynamic-load precedent exactly,
  including the same type validation).

Extracted the crawler manifest registry into a new shared
app/api/src/crawlerManifests.js so jobs.js and the renamed crawlerFiles.js
don't need to import from each other circularly (jobs.js previously built
this registry inline and imported csvUploads.js; the renamed file now also
needs the registry). jobs.js re-exports VALID_JOB_TYPES/validateCrawlerConfig
so scheduler.js and the two existing jobs.*.test.js files that import these
directly from jobs.js needed zero changes.

Fixes a previously-deferred issue for free: csvUploads.js's SCHEMA_HEADERS
was a third hardcoded copy of CSV column headers (missing ContextMembers.csv),
kept inline specifically because tools/csv-templates/ wasn't mirrored into
the Docker/node-launcher build pipelines. tools/crawlers/<type>/ already is
(proven in #354) — moved the real template files into tools/crawlers/csv/schema/
so the route can read real files instead of a third hand-maintained copy,
which also fixes the missing ContextMembers.csv as a byproduct.

Verified end-to-end against a real Docker build + live container: upload/
list/delete cycle, folder naming convention preserved exactly (csv-1), a
non-upload-supporting config type correctly rejected (400), job dispatch
correctly resolves the generalized folder path and gates on file presence,
config deletion correctly cleans up the folder, and the schema endpoint now
serves all 10 files (including the previously-missing ContextMembers.csv)
with correct label/required annotations sourced from the real files.
