+++ app/api/CLAUDE.md
@@ -11,6 +11,7 @@ Before writing any helper, utility, middleware, or route logic — search first.
 - `middleware/auth.js` — Entra ID JWT validation (v1 + v2 tokens)
 - `middleware/perfMetrics.js` — request timing + Server-Timing headers
 - `secrets/vault.js` — encrypted secret storage/retrieval for crawler credentials
+- `crawlerManifests.js` — the crawler manifest registry (`CRAWLER_MANIFESTS_DIR`, `VALID_JOB_TYPES`, `validateCrawlerConfig`), scanned once at startup from `tools/crawlers/*/crawler.json`; shared by `routes/jobs.js` and `routes/crawlerFiles.js` to avoid a circular import between them
 
 ## Always Test Locally Before Committing
